+++ Sources/PCCurrentUser.swift
@@ -679,10 +679,21 @@ public final class PCCurrentUser {
     }
 
     // TODO: Do I need to add a Last-Event-ID option here?
-    public func subscribeToRoom(room: PCRoom, roomDelegate: PCRoomDelegate, messageLimit: Int = 20) {
+    public func subscribeToRoom(
+        room: PCRoom,
+        roomDelegate: PCRoomDelegate,
+        messageLimit: Int = 20,
+        completionHandler: @escaping PCErrorCompletionHandler
+    ) {
         self.instance.logger.log(
             "About to subscribe to room \(room.debugDescription)",
             logLevel: .verbose
+        )
+
+        let completionHandler = steppedCompletionHandler(
+            steps: 3,
+            inner: completionHandler,
+            dispatchQueue: roomSubscriptionQueue
         )
 
         self.joinRoom(roomId: room.id) { innerRoom, err in
@@ -696,17 +707,25 @@ public final class PCCurrentUser {
 
             let messageSub = self.subscribeToRoomMessages(
                 room: roomToSubscribeTo,
-                roomDelegate: roomDelegate,
-                messageLimit: messageLimit
+                delegate: roomDelegate,
+                messageLimit: messageLimit,
+                completionHandler: completionHandler
             )
             let cursorSub = self.subscribeToRoomCursors(
                 room: roomToSubscribeTo,
-                delegate: roomDelegate
+                delegate: roomDelegate,
+                completionHandler: completionHandler
+            )
+            let membershipSub = self.subscribeToRoomMemberships(
+                room: roomToSubscribeTo,
+                delegate: roomDelegate,
+                completionHandler: completionHandler
             )
 
             room.subscription = PCRoomSubscription(
                 messageSubscription: messageSub,
                 cursorSubscription: cursorSub,
+                membershipSubscription: membershipSub,
                 delegate: roomDelegate
             )
         }
@@ -714,8 +733,9 @@ public final class PCCurrentUser {
 
     fileprivate func subscribeToRoomMessages(
         room: PCRoom,
-        roomDelegate: PCRoomDelegate,
-        messageLimit: Int
+        delegate: PCRoomDelegate,
+        messageLimit: Int,
+        completionHandler: @escaping PCErrorCompletionHandler
     ) -> PCMessageSubscription {
         let path = "/rooms/\(room.id)"
 
@@ -735,7 +755,7 @@ public final class PCCurrentUser {
         )
 
         let messageSubscription = PCMessageSubscription(
-            delegate: roomDelegate,
+            delegate: delegate,
             chatManagerDelegate: self.delegate,
             resumableSubscription: resumableSub,
             logger: self.instance.logger,
@@ -753,16 +773,22 @@ public final class PCCurrentUser {
         self.instance.subscribeWithResume(
             with: &resumableSub,
             using: subscribeRequest,
+            onOpen: { completionHandler(nil) },
             onEvent: { [unowned messageSubscription] eventID, headers, data in
                 messageSubscription.handleEvent(eventId: eventID, headers: headers, data: data)
-            }
-            // TODO: Should we be handling onError here somehow?
+            },
+            onError: completionHandler
+
         )
 
         return messageSubscription
     }
 
-    fileprivate func subscribeToRoomCursors(room: PCRoom, delegate: PCRoomDelegate) -> PCCursorSubscription {
+    fileprivate func subscribeToRoomCursors(
+        room: PCRoom,
+        delegate: PCRoomDelegate,
+        completionHandler: @escaping PCErrorCompletionHandler
+    ) -> PCCursorSubscription {
         let path = "/cursors/\(PCCursorType.read.rawValue)/rooms/\(room.id)"
 
         let subscribeRequest = PPRequestOptions(
@@ -781,10 +807,7 @@ public final class PCCurrentUser {
             cursorStore: cursorStore,
             connectionCoordinator: connectionCoordinator,
             logger: self.cursorsInstance.logger,
-            initialStateHandler: { err in
-                // TODO: Only consider the room subscription open when both the
-                // room subscription and cursor subscription have opened
-            }
+            initialStateHandler: completionHandler
         )
 
         self.cursorsInstance.subscribeWithResume(
@@ -792,11 +815,50 @@ public final class PCCurrentUser {
             using: subscribeRequest,
             onEvent: { [unowned cursorSubscription] eventID, headers, data in
                 cursorSubscription.handleEvent(eventId: eventID, headers: headers, data: data)
-            }
-            // TODO: Should we be handling onError here somehow?
+            },
+            onError: completionHandler
         )
 
         return cursorSubscription
+    }
+
+    fileprivate func subscribeToRoomMemberships(
+        room: PCRoom,
+        delegate: PCRoomDelegate,
+        completionHandler: @escaping PCErrorCompletionHandler
+    ) -> PCMembershipSubscription {
+        let path = "/rooms/\(room.id)/memberships"
+
+        let subscribeRequest = PPRequestOptions(
+            method: HTTPMethod.SUBSCRIBE.rawValue,
+            path: path
+        )
+
+        var resumableSub = PPResumableSubscription(
+            instance: self.instance,
+            requestOptions: subscribeRequest
+        )
+
+        let membershipSubscription = PCMembershipSubscription(
+            roomId: room.id,
+            delegate: delegate,
+            resumableSubscription: resumableSub,
+            userStore: self.userStore,
+            roomStore: self.roomStore,
+            logger: self.instance.logger,
+            initialStateHandler: completionHandler
+        )
+
+        self.instance.subscribeWithResume(
+            with: &resumableSub,
+            using: subscribeRequest,
+            onEvent: { [unowned membershipSubscription] eventID, headers, data in
+                membershipSubscription.handleEvent(eventId: eventID, headers: headers, data: data)
+            },
+            onError: completionHandler
+        )
+
+        return membershipSubscription
     }
 
     public func fetchMessagesFromRoom(
@@ -982,3 +1044,28 @@ public enum PCRoomMessageFetchDirection: String {
 public typealias PCErrorCompletionHandler = (Error?) -> Void
 public typealias PCRoomCompletionHandler = (PCRoom?, Error?) -> Void
 public typealias PCRoomsCompletionHandler = ([PCRoom]?, Error?) -> Void
+
+// Takes an `inner` completion handler of type `PCErrorCompletionHandler`,
+// returns another completion handler of the same type that calls the inner
+// completion handler after being called itself `steps` times. Returns the last
+// error, if any.
+func steppedCompletionHandler(
+    steps: Int,
+    inner: @escaping PCErrorCompletionHandler,
+    dispatchQueue: DispatchQueue
+) -> PCErrorCompletionHandler {
+    var error: Error?
+
+    let group = DispatchGroup()
+
+    for _ in 0..<steps {
+        group.enter()
+    }
+
+    group.notify(queue: dispatchQueue) { inner(error) }
+
+    return { err in
+        error = err
+        group.leave()
+    }
+}