+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `sendMessage` now requires a value for the `text` parameter
 - All `PCMessage`s must have a `text` property
+- `subscribeToRoom` now has a (required) completion handler of type
+  `PCErrorCompletionHandler`
 
 ## [0.8.4](https://github.com/pusher/chatkit-swift/compare/0.8.3...0.8.4) - 2018-05-26